requirements.txt
numpy<1.19.0
tensorflow-cpu==2.3.1
seaborn==0.11.0
